+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [unreleased]
 
+## [2.2.0]
+
 ### Added
 
 - Add opam support for the Dependency submission API.
@@ -530,7 +532,8 @@ that require it.
 
 - Initial release.
 
-[unreleased]: https://github.com/ocaml/setup-ocaml/compare/v2.1.8...HEAD
+[unreleased]: https://github.com/ocaml/setup-ocaml/compare/v2.2.0...HEAD
+[2.2.0]: https://github.com/ocaml/setup-ocaml/compare/v2.1.8...v2.2.0
 [2.1.8]: https://github.com/ocaml/setup-ocaml/compare/v2.1.7...v2.1.8
 [2.1.7]: https://github.com/ocaml/setup-ocaml/compare/v2.1.6...v2.1.7
 [2.1.6]: https://github.com/ocaml/setup-ocaml/compare/v2.1.5...v2.1.6